+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Drastically improve application logging
 * Add missing error handling to the `/summary` route, change HTTP error responses to return JSON
+* Add `-v` option which enables debug logging
 
 ## 2.3.2